+++ certificates/README.rst
@@ -1,0 +1,5 @@
+Goals:
+- An all in one binary to work with certificates (hence the current "heredoc" like variables in var-data.go)
+
+
+The web interface should be very simple to use and not expose many complexities of x509 certificates. But it should also be strict and not allow for misconfigurations or other mistakes.
